requirements.txt
streamlit
google-cloud-firestore
protobuf==3.20.1
